feat: Implement CouponAdminForm for Coupon model in admin panel
Created the 'CouponAdminForm' to enhance the admin interface for the 'Coupon' model. This form includes additional validation to ensure that both start and end dates, as well as the number available and number used, are provided together. This validation helps maintain consistency in data entry for coupon instances.

feat: Implement CouponRuleAdminFormSet for CouponRule model in admin panel

Introduced the 'CouponRuleAdminFormSet' to manage forms for the 'CouponRule' model within the admin panel. This formset incorporates validation to ensure that the start prices are in ascending order and that a bigger discount corresponds to a bigger start price. These measures contribute to maintaining the integrity and coherence of discount rules in the system.