feat: Implement OrderItemAdminFormSet for OrderItem model in admin panel
Introduced the 'OrderItemAdminFormSet' to manage forms for the 'OrderItem' model within the admin panel. This formset includes validation logic that dynamically calculates and sets the price, discount, and discount price based on the selected product. ensuring data integrity and consistency in the admin interface.

feat: Implement OrderItemAdminForm for OrderItem model in admin panel

Created the 'OrderItemAdminForm' to enhance the admin interface for the 'OrderItem' model. This form includes validation logic to prevent the association of items with completed orders, ensuring data consistency and accuracy.