feat: Introduce 'coupon' and 'coupon_price' fields in Order model
Added the 'coupon' field to associate orders with a coupon, allowing for discounts and promotions. The field is set to allow null values and uses the 'SET_NULL' on delete behavior.

Also introduced the 'coupon_price' field that stores the amount by which the total price of the order is to be reduced when a coupon is applied.  This field enhances the representation of coupon-related pricing in the Order model.

refactor: Rename methods in OrderItem model for consistency

Renamed the methods in the 'OrderItem' model for better consistency and clarity:

1. Renamed 'get_total_no_discount' to 'get_total_no_discount_item' to ensure a more descriptive and consistent naming convention.
2. Renamed 'get_total_with_discount' to 'get_total_with_discount_item' for improved method naming consistency.
3. Renamed 'get_total_profit' to 'get_total_profit_item' to maintain naming coherence.
4. Renamed 'get_total' to 'get_total_item' to provide a standardized and clear method name.

These changes enhance the readability and maintainability of the 'OrderItem' model methods.