feat: Introduce 'cart_view' for rendering and processing the shopping cart page
Added a view function, 'cart_view,' to handle the rendering and processing of the shopping cart page. The view incorporates a CouponForm to handle coupon submissions, validating and applying discounts to the cart when applicable. For authenticated users, the view interacts with the 'Order' model, while for anonymous users, it utilizes the 'Cart' class.

 This enhances the overall shopping cart experience by providing a dynamic and responsive interface for users to view and manage their selected items.

feat: Introduce 'cart/' URL pattern for the 'cart_view' in the store app

Added a new URL pattern 'cart/' in the store app's 'urls.py' file to map to the 'cart_view' function. This URL pattern directs users to the shopping cart page, enhancing the navigation and user experience by providing a dedicated route for viewing and managing the contents of the shopping cart.