feat: Introduce signal in 'product_signals.py' to automatically set ProductColorAndSizeValue as inactive
Implemented a signal in 'product_signals.py' – `set_product_color_size_value_as_inactive` – triggered before saving a ProductColorAndSizeValue instance. This signal checks if the inventory of the color-size value is zero or negative, and if so, marks the instance as inactive. This automatic marking helps streamline the handling of inactive color-size values, particularly in scenarios where the inventory reaches or falls below zero during updates.

This signal works in tandem with the 'order_signals.py' signal `delete_inactive_product_color_size_in_order_itme` to ensure the proper removal of associated OrderItems when color-size values become inactive or out of stock.