feat: Add `get_absolute_url` method to Category model
Implemented the `get_absolute_url` method in the Category model to generate the absolute URL for the category. This enhancement facilitates easier navigation and linking to category-specific pages throughout the application.

feat: Add `ActiveProductColorAndSizeValueManager` to ProductColorAndSizeValue model

Introduced a custom manager, `ActiveProductColorAndSizeValueManager`, to the `ProductColorAndSizeValue` model. This manager filters queryset based on conditions to include only active records with available inventory or unlimited stock. The `active_objs` manager is set to use this custom manager for retrieving active records. This implementation improves the handling of active product color and size values.