refactor: Improve `calculate_coupon_price` method in Order model
Enhanced the `calculate_coupon_price` method in the `Order` model to improve readability and messaging.

The refactored method now includes an additional argument `success_message` to control whether success messages are displayed. If `success_message` is `True`, a success message will be shown upon successfully applying the coupon. If unsuccessful, the method sets the `coupon` field to `None` and updates the `coupon_price` to 0, ensuring consistent behavior and improved user feedback.

These changes enhance the clarity and flexibility of the method.

feat: Add `get_absolute_url` method to Order model

Introduced the `get_absolute_url` method in the `Order` model, enabling the generation of the absolute URL for an order instance. This method enhances navigation and provides a direct link to view the details of a specific order.